Do not create database connections in the number of input files
It does not make sense to create database connection generators
in the number of input tables since, well, DB connections are
independent of a given table.
Also the possibility of multiple database connections is not
really relevant, since I do not know of an algorithm that
makes use of querying multiple distinct database instances.